+++ src/Annotations/Mappings/Number/ScaledFloat.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Nadia\ElasticSearchODM\Annotations\Mappings\Number;
+
+/**
+ * @Annotation
+ * @Target({"ANNOTATION"})
+ */
+final class ScaledFloat
+{
+    public $coerce = true;
+
+    public $boost = 1.0;
+
+    public $doc_values = true;
+
+    public $ignore_malformed = false;
+
+    public $include_in_all = false;
+
+    public $index = true;
+
+    public $null_value;
+
+    public $store = false;
+
+    public $scaling_factor = 1;
+}